checkpointing the beginning of async workers, starting with acq. not functional yet.
