+++ ifcb2/workflow/__init__.py
@@ -1,0 +1,8 @@
+"""IFCB workflow components,
+including workers for
+- acquisition
+- accession"""
+
+WILD_PRODUCT='wild'
+RAW_PRODUCT='raw'
+ACCESSION_ROLE='accession'
