+++ workflow/orm.py
@@ -230,7 +230,6 @@ class Products(object):
         - have any dependencies (in other words, not "root" products")
         - have any dependents, all of which are in 'dep_state'
         default is to find available products that no unavailable products depend on"""
-        self.session.begin_nested()
         for product in self.session.query(Product).\
             filter(Product.state==state).\
             filter(Product.depends_on.any()).\
@@ -246,7 +245,6 @@ class Products(object):
             raise ValueError('state and new_state are both %s' % state)
         now = utcdtnow()
         n = 0
-        self.session.begin_nested()
         for p in self.session.query(Product).\
             filter(Product.expires.isnot(None)).\
             filter(now > Product.expires).\
